+++ src/websocketadapter/DefaultWebSocketAdapter.ts
@@ -1,4 +1,4 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 import Logger from '../logger/Logger';
@@ -29,8 +29,8 @@ export default class DefaultWebSocketAdapter implements WebSocketAdapter {
     }
   }
 
-  close(): void {
-    this.connection.close();
+  close(code?: number, reason?: string): void {
+    this.connection.close(code, reason);
   }
 
   destroy(): void {