+++ test/signalingclient/SignalingClientConnectionRequest.test.ts
@@ -1,4 +1,4 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 import * as chai from 'chai';
@@ -7,7 +7,8 @@ import SignalingClientConnectionRequest from '../../src/signalingclient/Signalin
 
 describe('SignalingClientConnectionRequest', () => {
   const expect: Chai.ExpectStatic = chai.expect;
-  const fullSignalingURL = 'signalingurl/control/conferenceid?X-Chime-Control-Protocol-Version=3';
+  const fullSignalingURL =
+    'signalingurl/control/conferenceid?X-Chime-Control-Protocol-Version=3&X-Amzn-Chime-Send-Close-On-Error=1';
   describe('url', () => {
     it('use the url as given if it is already fully constructed', () => {
       const request = new SignalingClientConnectionRequest(