+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update Travis config to improve PR build speed
 - Disable saucelab capabilities
+- Use credentials sent via signaling connection JOIN_ACK to improve audio-video startup time.
 
 ### Removed